backend/requirements.txt
# Python 3.10+
uvicorn==0.27.0
fastapi==0.109.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4
langchain-community==0.0.16
langchain==0.1.5
tenacity==8.2.3
PyYAML==6.0.1
python-docx==1.1.0
pdfplumber==0.10.3
Pillow==10.2.0
element-path==4.2.0
